# backend/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/library/tracks/{trackId}/tags": {
      "post": {
        "operationId": "LibraryController_addTagToTrack",
        "parameters": [
          {
            "name": "trackId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddTagToTrackDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Tag added to track"
          },
          "201": {
            "description": ""
          },
          "404": {
            "description": "Track or tag not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Add a tag to a track",
        "tags": [
          "library"
        ]
      }
    },
    "/library/tags": {
      "post": {
        "operationId": "LibraryController_createTag",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTagDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Tag created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TagResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a new tag",
        "tags": [
          "library"
        ]
      },
      "get": {
        "operationId": "LibraryController_getTags",
        "parameters": [],
        "responses": {
          "200": {
            "description": "List of tags",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TagResponseDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get all user tags",
        "tags": [
          "library"
        ]
      }
    },
    "/library/tags/{tagId}": {
      "delete": {
        "operationId": "LibraryController_deleteTag",
        "parameters": [
          {
            "name": "tagId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "204": {
            "description": "Tag deleted"
          },
          "404": {
            "description": "Tag not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete a tag",
        "tags": [
          "library"
        ]
      },
      "put": {
        "operationId": "LibraryController_updateTag",
        "parameters": [
          {
            "name": "tagId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTagDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Tag updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TagResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Tag not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update a tag",
        "tags": [
          "library"
        ]
      }
    },
    "/library/albums": {
      "get": {
        "operationId": "LibraryController_getAlbums",
        "parameters": [
          {
            "name": "genres",
            "required": false,
            "in": "query",
            "description": "Filter by artist genres (albums matching any of the genres)",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "description": "Page size",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 24,
              "type": "number"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Search query for album name or artist name",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sortBy",
            "required": false,
            "in": "query",
            "description": "Sort field",
            "schema": {
              "type": "string",
              "enum": [
                "artist",
                "avgRating",
                "lastPlayed",
                "name",
                "totalPlayCount",
                "trackCount"
              ]
            }
          },
          {
            "name": "sortOrder",
            "required": false,
            "in": "query",
            "description": "Sort order",
            "schema": {
              "default": "asc",
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated list of albums",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedAlbumsDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get all albums in user library",
        "tags": [
          "library"
        ]
      }
    },
    "/library/albums/{artist}/{album}/tracks": {
      "get": {
        "operationId": "LibraryController_getAlbumTracks",
        "parameters": [
          {
            "name": "artist",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "album",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of tracks from the album",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlbumTracksResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get tracks from a specific album",
        "tags": [
          "library"
        ]
      }
    },
    "/library/artists": {
      "get": {
        "operationId": "LibraryController_getArtists",
        "parameters": [
          {
            "name": "genres",
            "required": false,
            "in": "query",
            "description": "Filter by genres (artists matching any of the genres)",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "description": "Page size",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 24,
              "type": "number"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Search query for artist name",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sortBy",
            "required": false,
            "in": "query",
            "description": "Sort field",
            "schema": {
              "type": "string",
              "enum": [
                "albumCount",
                "avgRating",
                "lastPlayed",
                "name",
                "totalPlayCount",
                "trackCount"
              ]
            }
          },
          {
            "name": "sortOrder",
            "required": false,
            "in": "query",
            "description": "Sort order",
            "schema": {
              "default": "asc",
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated list of artists",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedArtistsDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get all artists in user library",
        "tags": [
          "library"
        ]
      }
    },
    "/library/artists/{artist}/tracks": {
      "get": {
        "operationId": "LibraryController_getArtistTracks",
        "parameters": [
          {
            "name": "artist",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of tracks from the artist",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArtistTracksResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get tracks from a specific artist",
        "tags": [
          "library"
        ]
      }
    },
    "/library/genres": {
      "get": {
        "operationId": "LibraryController_getGenres",
        "parameters": [],
        "responses": {
          "200": {
            "description": "List of unique genres",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get all unique genres in user library",
        "tags": [
          "library"
        ]
      }
    },
    "/library/play-history": {
      "get": {
        "operationId": "LibraryController_getPlayHistory",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "description": "Page size",
            "schema": {
              "minimum": 1,
              "maximum": 1000,
              "default": 50,
              "type": "number"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Search query for title, artist or album",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "trackId",
            "required": false,
            "in": "query",
            "description": "Filter by track ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated list of play history",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedPlayHistoryDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get user play history",
        "tags": [
          "library"
        ]
      }
    },
    "/library/sync/count": {
      "get": {
        "operationId": "LibraryController_getSyncItemCounts",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Item counts retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SyncItemCountsDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Pre-count items to be synced (tracks, albums, playlists)",
        "tags": [
          "library"
        ]
      }
    },
...
  },
  "info": {
    "title": "Libify API",
    "description": "Spotify library management API",
    "version": "0.0.1",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "AddTagToTrackDto": {
        "type": "object",
        "properties": {
          "tagId": {
            "type": "string",
            "format": "uuid",
            "description": "Tag ID to add to the track"
          }
        },
        "required": [
          "tagId"
        ]
      },
      "CreateTagDto": {
        "type": "object",
        "properties": {
          "color": {
            "type": "string",
            "pattern": "^#?([0-9A-F]{3}|[0-9A-F]{4}|[0-9A-F]{6}|[0-9A-F]{8})$",
            "description": "Tag color in hex format"
          },
          "name": {
            "type": "string",
            "description": "Tag name"
          }
        },
        "required": [
          "name"
        ]
      },
      "TagResponseDto": {
        "type": "object",
        "properties": {
          "color": {
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          }
        },
        "required": [
          "createdAt",
          "id",
          "name"
        ]
      },
...
      "UpdateTagDto": {
        "type": "object",
        "properties": {
          "color": {
            "type": "string",
            "pattern": "^#?([0-9A-F]{3}|[0-9A-F]{4}|[0-9A-F]{6}|[0-9A-F]{8})$",
            "description": "Tag color in hex format"
          },
          "name": {
            "type": "string",
            "description": "Tag name"
          }
        }
      },
...
    }
  }
}